--- src/main/java/com/team3/otboo/common/event/EventPayload.java
@@ -1,5 +1,0 @@
-package com.team3.otboo.common.event;
-
-public interface EventPayload {
-
-}
