+++ src/main/java/com/team3/otboo/domain/feed/service/FeedService.java
@@ -11,7 +11,7 @@ import com.team3.otboo.domain.feed.repository.FeedRepositoryQueryDSL;
 import com.team3.otboo.domain.feed.service.request.FeedCreateRequest;
 import com.team3.otboo.domain.feed.service.request.FeedListRequest;
 import com.team3.otboo.domain.feed.service.request.FeedUpdateRequest;
-import com.team3.otboo.common.outboxMessageRelay.OutboxEventPublisher;
+import com.team3.otboo.domain.hot.common.outboxMessageRelay.OutboxEventPublisher;
 import jakarta.persistence.EntityNotFoundException;
 import java.util.List;
 import java.util.UUID;
