+++ src/main/java/com/team3/otboo/domain/hot/common/event/EventPayload.java
@@ -1,0 +1,5 @@
+package com.team3.otboo.domain.hot.common.event;
+
+public interface EventPayload {
+
+}
