+++ src/main/java/com/team3/otboo/domain/hot/common/outboxMessageRelay/MessageRelayConfig.java
@@ -1,4 +1,4 @@
-package com.team3.otboo.common.outboxMessageRelay;
+package com.team3.otboo.domain.hot.common.outboxMessageRelay;
 
 import java.util.HashMap;
 import java.util.Map;
@@ -17,7 +17,7 @@ import org.springframework.scheduling.annotation.EnableScheduling;
 import org.springframework.scheduling.concurrent.ThreadPoolTaskExecutor;
 
 @Configuration
-@ComponentScan("com.team3.otboo.common.outboxMessageRelay")
+@ComponentScan("com.team3.otboo.domain.hot.common.outboxMessageRelay")
 @EnableAsync // 비동기 처리 .
 @EnableScheduling
 public class MessageRelayConfig {
