+++ src/main/java/com/team3/otboo/domain/hot/common/outboxMessageRelay/OutboxRepository.java
@@ -1,4 +1,4 @@
-package com.team3.otboo.common.outboxMessageRelay;
+package com.team3.otboo.domain.hot.common.outboxMessageRelay;
 
 import java.time.Instant;
 import java.util.List;
