+++ src/main/java/com/team3/otboo/domain/hot/service/HotFeedScoreUpdater.java
@@ -2,8 +2,8 @@ package com.team3.otboo.domain.hot.service;
 
 import com.team3.otboo.domain.feed.entity.Feed;
 import com.team3.otboo.domain.feed.service.FeedService;
-import com.team3.otboo.common.event.Event;
-import com.team3.otboo.common.event.EventPayload;
+import com.team3.otboo.domain.hot.common.event.Event;
+import com.team3.otboo.domain.hot.common.event.EventPayload;
 import com.team3.otboo.domain.hot.repository.HotFeedListRepository;
 import com.team3.otboo.domain.hot.service.eventhandler.EventHandler;
 import java.time.Duration;
