+++ src/main/java/com/team3/otboo/domain/hot/service/HotFeedService.java
@@ -3,9 +3,9 @@ package com.team3.otboo.domain.hot.service;
 import com.team3.otboo.domain.feed.dto.FeedDto;
 import com.team3.otboo.domain.feed.mapper.FeedDtoAssembler;
 import com.team3.otboo.domain.hot.client.FeedClient;
-import com.team3.otboo.common.event.Event;
-import com.team3.otboo.common.event.EventPayload;
-import com.team3.otboo.common.event.EventType;
+import com.team3.otboo.domain.hot.common.event.Event;
+import com.team3.otboo.domain.hot.common.event.EventPayload;
+import com.team3.otboo.domain.hot.common.event.EventType;
 import com.team3.otboo.domain.hot.repository.HotFeedListRepository;
 import com.team3.otboo.domain.hot.service.eventhandler.EventHandler;
 import java.util.List;
