+++ src/main/java/com/team3/otboo/domain/hot/service/eventhandler/CommentCreatedEventHandler.java
@@ -1,8 +1,8 @@
 package com.team3.otboo.domain.hot.service.eventhandler;
 
-import com.team3.otboo.common.event.Event;
-import com.team3.otboo.common.event.EventType;
-import com.team3.otboo.common.event.payload.CommentCreatedEventPayload;
+import com.team3.otboo.domain.hot.common.event.Event;
+import com.team3.otboo.domain.hot.common.event.EventType;
+import com.team3.otboo.domain.hot.common.event.payload.CommentCreatedEventPayload;
 import java.util.UUID;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
